Add thermal reading data view
